+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support displaying RTF content (including encapsulated HTML) from a TNEF attachment
+- Fix size of Mailvelope iframe for PGP-inlined mail, again (#8126)
 
 ## Release 1.5-rc